Round 1: resolve general compiler errors which is non-related to the dependency of appache/xerec 3rd party library

+++ Source/CompoundNode.cpp
@@ -33,7 +33,7 @@ ILxNode* CompoundNode::getChild (int nIndex) const
 
 int CompoundNode::numberChildren () const
 {
-	return m_arrChildren.size();
+	return (int)m_arrChildren.size();
 }
 
 ILxNode* CompoundNode::removeChild (int nIndex)
